+++ packages/eyes-cypress/src/browser/commands.js
@@ -1,46 +1,41 @@
 /* global Cypress,cy,window,before,after,navigator */
 'use strict';
 const poll = require('./poll');
-const makeSend = require('./makeSend');
 const processPage = require('@applitools/dom-snapshot/dist/processPageCjs');
 const domSnapshotOptions = {dontFetchResources: Cypress.config('eyesDisableBrowserFetching')};
-const send = makeSend(Cypress.config('eyesPort'), window.fetch);
-const makeSendRequest = require('./sendRequest');
 const makeEyesCheckWindow = require('./eyesCheckWindow');
 const makeHandleCypressViewport = require('./makeHandleCypressViewport');
-const sendRequest = makeSendRequest(send);
-const eyesCheckWindow = makeEyesCheckWindow({sendRequest, processPage, domSnapshotOptions});
 const handleCypressViewport = makeHandleCypressViewport({cy});
+const makeSend = require('./makeSend');
+const send = makeSend(Cypress.config('eyesPort'), window.fetch);
+const makeSendRequest = require('./sendRequest');
+const sendRequest = makeSendRequest(send);
+
+const eyesCheckWindow = makeEyesCheckWindow({sendRequest, processPage, domSnapshotOptions});
 
 function getGlobalConfigProperty(prop) {
   const property = Cypress.config(prop);
   const shouldParse = ['eyesBrowser', 'eyesLayoutBreakpoints'];
   return property ? (shouldParse.includes(prop) ? JSON.parse(property) : property) : undefined;
 }
-
-if (!Cypress.config('eyesIsDisabled')) {
-  const batchEnd = poll(({timeout}) => {
-    return sendRequest({command: 'batchEnd', data: {timeout}});
+if (
+  !getGlobalConfigProperty('eyesIsDisabled') &&
+  (getGlobalConfigProperty('eyesLegacyHooks') || getGlobalConfigProperty('isInteractive'))
+) {
+  const batchEnd = poll(() => {
+    return sendRequest({command: 'batchEnd'});
   });
 
   before(() => {
-    const userAgent = navigator.userAgent;
-    const viewport = {
-      width: getGlobalConfigProperty('viewportWidth'),
-      height: getGlobalConfigProperty('viewportHeight'),
-    };
-    let browser = getGlobalConfigProperty('eyesBrowser');
-    handleCypressViewport(browser).then({timeout: 86400000}, () =>
-      sendRequest({
-        command: 'batchStart',
-        data: {viewport, userAgent, isInteractive: getGlobalConfigProperty('isInteractive')},
-      }),
-    );
+    sendRequest({
+      command: 'batchStart',
+      data: {isInteractive: getGlobalConfigProperty('isInteractive')},
+    });
   });
 
   after(() => {
     cy.then({timeout: 86400000}, () => {
-      return batchEnd({timeout: getGlobalConfigProperty('eyesTimeout')}).catch(e => {
+      return batchEnd().catch(e => {
         if (!!getGlobalConfigProperty('eyesFailCypressOnDiff')) {
           throw e;
         }
@@ -54,8 +49,18 @@ let isCurrentTestDisabled;
 Cypress.Commands.add('eyesOpen', function(args = {}) {
   Cypress.config('eyesOpenArgs', args);
   Cypress.log({name: 'Eyes: open'});
+  const userAgent = navigator.userAgent;
   const {title: testName} = this.currentTest || this.test;
-  const {browser, isDisabled} = args;
+  const {browser: eyesOpenBrowser, isDisabled} = args;
+  const globalBrowser = getGlobalConfigProperty('eyesBrowser');
+  const defaultBrowser = {
+    width: getGlobalConfigProperty('viewportWidth'),
+    height: getGlobalConfigProperty('viewportHeight'),
+    name: 'chrome',
+  };
+
+  const browser =
+    validateBrowser(eyesOpenBrowser) || validateBrowser(globalBrowser) || defaultBrowser;
 
   if (Cypress.config('eyesIsDisabled') && isDisabled === false) {
     throw new Error(
@@ -73,15 +78,10 @@ Cypress.Commands.add('eyesOpen', function(args = {}) {
     }
   }
 
-  function fillDefaultBrowserName(browser) {
-    if (!browser.name && !browser.iosDeviceInfo && !browser.chromeEmulationInfo) {
-      browser.name = 'chrome';
-    }
-  }
   return handleCypressViewport(browser).then({timeout: 15000}, () =>
     sendRequest({
       command: 'open',
-      data: Object.assign({testName}, args),
+      data: Object.assign({testName}, args, {browser, userAgent}),
     }),
   );
 });
@@ -123,3 +123,16 @@ Cypress.Commands.add('eyesClose', () => {
   }
   return sendRequest({command: 'close'});
 });
+
+function fillDefaultBrowserName(browser) {
+  if (!browser.name && !browser.iosDeviceInfo && !browser.chromeEmulationInfo) {
+    browser.name = 'chrome';
+  }
+}
+
+function validateBrowser(browser) {
+  if (!browser) return false;
+  if (Array.isArray(browser)) return browser.length ? browser : false;
+  if (Object.keys(browser).length === 0) return false;
+  return browser;
+}