+++ packages/eyes-cypress/src/plugin/handlers.js
@@ -4,28 +4,18 @@ const pollingHandler = require('./pollingHandler');
 const makeWaitForBatch = require('./waitForBatch');
 const makeHandleBatchResultsFile = require('./makeHandleBatchResultsFile');
 const {GeneralUtils} = require('@applitools/visual-grid-client');
-
-const TIMEOUT_MSG = timeout =>
-  `Eyes-Cypress timed out after ${timeout}ms. The default timeout is 2 minutes. It's possible to increase this timeout by setting a the value of 'eyesTimeout' in Cypress configuration, e.g. for 3 minutes: Cypress.config('eyesTimeout', 180000)`;
+const runningTests = require('./runningTests');
 
 function makeHandlers({
-  makeVisualGridClient,
   config = {},
+  visualGridClient,
   logger = console,
   processCloseAndAbort,
   getErrorsAndDiffs,
   errorDigest,
 }) {
   logger.log('[handlers] creating handlers with the following config:', config);
-  let openEyes,
-    pollBatchEnd,
-    checkWindow,
-    close,
-    resources,
-    openErr,
-    getEmulatedDevicesSizes,
-    getIosDevicesSizes;
-  let runningTests = [];
+  let pollBatchEnd, checkWindow, close, resources, openErr;
 
   return {
     open: async args => {
@@ -33,7 +23,7 @@ function makeHandlers({
         logger.log(`[handlers] open: close=${typeof close}, args=`, args);
         args.accessibilitySettings = args.accessibilityValidation;
         delete args.accessibilityValidation;
-        const eyes = await openEyes(args);
+        const eyes = await visualGridClient.openEyes(args);
         const runningTest = {
           abort: eyes.abort,
           closePromise: undefined,
@@ -41,7 +31,7 @@ function makeHandlers({
         checkWindow = eyes.checkWindow;
         close = makeClose(eyes.close, runningTest);
         resources = {};
-        runningTests.push(runningTest);
+        runningTests.add(runningTest);
         logger.log('[handlers] open finished');
         return eyes;
       } catch (err) {
@@ -52,29 +42,11 @@ function makeHandlers({
     },
     batchStart: data => {
       logger.log('[handlers] batchStart with data', data);
-      runningTests = [];
-      const extraConfig = {};
-      if (
-        GeneralUtils.getPropertyByPath(data, 'viewport.height') &&
-        GeneralUtils.getPropertyByPath(data, 'viewport.width')
-      ) {
-        extraConfig.browser = data.viewport;
-      }
-      if (GeneralUtils.getPropertyByPath(data, 'userAgent')) {
-        extraConfig.userAgent = data.userAgent;
-      }
-
-      const client = makeVisualGridClient(
-        Object.assign(extraConfig, config, {
-          logger: (logger.extend && logger.extend('vgc')) || logger,
-        }),
-      );
-      openEyes = client.openEyes;
-      getEmulatedDevicesSizes = client.getEmulatedDevicesSizes;
-      getIosDevicesSizes = client.getIosDevicesSizes;
+      runningTests.reset();
+      const {testConcurrency} = config;
       const waitForBatch = makeWaitForBatch({
         logger: (logger.extend && logger.extend('waitForBatch')) || logger,
-        concurrency: config.concurrency,
+        testConcurrency,
         processCloseAndAbort,
         getErrorsAndDiffs,
         errorDigest,
@@ -82,16 +54,15 @@ function makeHandlers({
         handleBatchResultsFile: makeHandleBatchResultsFile(config),
       });
       pollBatchEnd = pollingHandler(
-        waitForBatch.bind(null, runningTests, client.closeBatch),
-        TIMEOUT_MSG,
+        waitForBatch.bind(null, runningTests.tests, visualGridClient.closeBatch),
       );
-      return client;
+      return visualGridClient;
     },
-    getIosDevicesSizes: () => getIosDevicesSizes(),
-    getEmulatedDevicesSizes: () => getEmulatedDevicesSizes(),
-    batchEnd: async ({timeout} = {}) => {
-      logger.log(`[handlers] batchEnd, timeout=${timeout}`);
-      return await pollBatchEnd({timeout});
+    getIosDevicesSizes: () => visualGridClient.getIosDevicesSizes(),
+    getEmulatedDevicesSizes: () => visualGridClient.getEmulatedDevicesSizes(),
+    batchEnd: async () => {
+      logger.log(`[handlers] batchEnd`);
+      return await pollBatchEnd();
     },
 
     putResource: (id, buffer) => {
@@ -227,4 +198,3 @@ function makeHandlers({
 }
 
 module.exports = makeHandlers;
-module.exports.TIMEOUT_MSG = TIMEOUT_MSG;