+++ packages/eyes-storybook/src/defaultConfig.js
@@ -1,7 +1,6 @@
 'use strict';
 
 module.exports = {
-  concurrency: 10,
   storybookPort: 9000,
   storybookHost: 'localhost',
   storybookConfigDir: '.storybook',
@@ -16,4 +15,5 @@ module.exports = {
   exitcode: true,
   readStoriesTimeout: 60000,
   reloadPagePerStory: false,
+  include: undefined,
 };