+++ packages/eyes-storybook/src/generateConfig.js
@@ -38,8 +38,12 @@ function generateConfig({argv = {}, defaultConfig = {}, externalConfigParams = [
     result.showLogs = true;
   }
 
-  if (result.storyDataGap === undefined && result.concurrency !== undefined) {
-    result.storyDataGap = result.concurrency;
+  if (!result.testConcurrency && !result.concurrency) {
+    result.testConcurrency = 5; // TODO require from core
+  }
+
+  if (result.storyDataGap === undefined) {
+    result.storyDataGap = result.testConcurrency;
   }
   return result;
 }