+++ packages/eyes-storybook/test/it/eyesStorybook.it.test.js
@@ -1,4 +1,4 @@
-const {describe, it, before, after} = require('mocha');
+const {describe, it, before, after, beforeEach, afterEach} = require('mocha');
 const flatten = require('lodash.flatten');
 const {expect} = require('chai');
 const testStorybook = require('../util/testStorybook');
@@ -7,6 +7,7 @@ const testServer = require('@applitools/sdk-shared/src/run-test-server');
 const fakeEyesServer = require('../util/fakeEyesServer');
 const eyesStorybook = require('../../src/eyesStorybook');
 const generateConfig = require('../../src/generateConfig');
+const defaultConfig = require('../../src/defaultConfig');
 const {configParams: externalConfigParams} = require('@applitools/visual-grid-client');
 const {makeTiming} = require('@applitools/monitoring-commons');
 const logger = require('../util/testLogger');
@@ -32,12 +33,12 @@ describe('eyesStorybook', () => {
   });
 
   let serverUrl, closeEyesServer;
-  before(async () => {
+  beforeEach(async () => {
     const {port, close} = await fakeEyesServer();
     closeEyesServer = close;
     serverUrl = `http://localhost:${port}`;
   });
-  after(async () => {
+  afterEach(async () => {
     await closeEyesServer();
   });
 
@@ -71,6 +72,7 @@ describe('eyesStorybook', () => {
       },
       {name: 'Button with-space yes-indeed/nested with-space yes: b yes-a b', isPassed: true},
       {name: 'Button with-space yes-indeed: a yes-a b', isPassed: true},
+      {name: 'Button: background color', isPassed: true},
       {name: 'Button: with some emoji', isPassed: true},
       {name: 'Button: with text', isPassed: true},
       {name: 'Image: image', isPassed: true},
@@ -120,6 +122,7 @@ describe('eyesStorybook', () => {
       'Button with-space yes-indeed: a yes-a b',
       'Button with-space yes-indeed/nested with-space yes: b yes-a b',
       'Button with-space yes-indeed/nested with-space yes/nested again-yes a: c yes-a b',
+      'Button: background color',
       'SOME section|Nested/Component: story 1.1',
       'SOME section|Nested/Component: story 1.2',
       'Wow|one with-space yes-indeed/nested with-space yes/nested again-yes a: c yes-a b',
@@ -134,6 +137,7 @@ describe('eyesStorybook', () => {
 
     expect(results.map(e => e.title).sort()).to.eql(expectedTitles.sort());
     results = flatten(results.map(r => r.resultsOrErr));
+
     expect(results.some(x => x instanceof Error)).to.be.false;
     expect(results).to.have.length(expectedResults.length);
 
@@ -145,7 +149,9 @@ describe('eyesStorybook', () => {
       const session = await fetch(sessionUrl).then(r => r.json());
       const {scenarioIdOrName} = session.startInfo;
       const [componentName, state] = scenarioIdOrName.split(':').map(s => s.trim());
+
       expect(session.startInfo.defaultMatchSettings.ignoreDisplacements).to.be.true;
+
       expect(session.startInfo.properties).to.eql([
         {name: 'Component name', value: componentName},
         {name: 'State', value: state.replace(/ \[.+\]$/, '')}, // strip off variation
@@ -171,6 +177,7 @@ describe('eyesStorybook', () => {
           coordinatesType: 'SCREENSHOT_AS_IS',
         },
       ]);
+
       expect(imageMatchSettings.floating).to.eql(floating);
       expect(imageMatchSettings.accessibility).to.eql(accessibility);
     }
@@ -183,8 +190,161 @@ describe('eyesStorybook', () => {
 
     expect(getEvents().join('')).to.equal(`- Reading stories
 ✔ Reading stories
-- Done 0 stories out of 19
-✔ Done 19 stories out of 19
+- Done 0 stories out of 20
+✔ Done 20 stories out of 20
 `);
+  });
+
+  it('enforces default concurrency', async () => {
+    const {stream} = testStream();
+    const configPath = path.resolve(__dirname, '../fixtures/applitools.config.js');
+    const config = generateConfig({argv: {conf: configPath}, defaultConfig, externalConfigParams});
+    await eyesStorybook({
+      config: {
+        ...config,
+        serverUrl,
+        storybookUrl: 'http://localhost:9001',
+      },
+      logger,
+      performance,
+      timeItAsync,
+      outputStream: stream,
+    });
+
+    const {maxRunning} = await fetch(`${serverUrl}/api/usage`).then(r => r.json());
+    expect(maxRunning).to.equal(5); // TODO require from core
+  });
+
+  it('enforces testConcurrency', async () => {
+    const {stream} = testStream();
+    const configPath = path.resolve(__dirname, '../fixtures/applitools.config.js');
+    const config = generateConfig({argv: {conf: configPath}, defaultConfig, externalConfigParams});
+    await eyesStorybook({
+      config: {
+        ...config,
+        serverUrl,
+        storybookUrl: 'http://localhost:9001',
+        testConcurrency: 3,
+      },
+      logger,
+      performance,
+      timeItAsync,
+      outputStream: stream,
+    });
+
+    const {maxRunning} = await fetch(`${serverUrl}/api/usage`).then(r => r.json());
+    expect(maxRunning).to.equal(3);
+  });
+
+  it('enforces testConcurrency over legacy concurrency', async () => {
+    const {stream} = testStream();
+    const configPath = path.resolve(__dirname, '../fixtures/applitools.config.js');
+    const config = generateConfig({argv: {conf: configPath}, defaultConfig, externalConfigParams});
+    await eyesStorybook({
+      config: {
+        ...config,
+        serverUrl,
+        storybookUrl: 'http://localhost:9001',
+        testConcurrency: 3,
+        concurrency: 20,
+      },
+      logger,
+      performance,
+      timeItAsync,
+      outputStream: stream,
+    });
+
+    const {maxRunning} = await fetch(`${serverUrl}/api/usage`).then(r => r.json());
+    expect(maxRunning).to.equal(3);
+  });
+
+  it('enforces legacy concurrency', async () => {
+    const {stream} = testStream();
+    const configPath = path.resolve(
+      __dirname,
+      '../fixtures/applitools-legacy-concurrency.config.js',
+    );
+    const config = generateConfig({argv: {conf: configPath}, defaultConfig, externalConfigParams});
+    await eyesStorybook({
+      config: {
+        ...config,
+        storybookUrl: 'http://localhost:9001',
+        serverUrl,
+      },
+      logger,
+      performance,
+      timeItAsync,
+      outputStream: stream,
+    });
+
+    const {maxRunning} = await fetch(`${serverUrl}/api/usage`).then(r => r.json());
+    expect(maxRunning).to.equal(10);
+  });
+
+  it('sends parentBranchBaselineSavedBefore when branchName and parentBranchName are specified, and there is a merge-base time for them', async () => {
+    const {stream} = testStream();
+    const configPath = path.resolve(
+      __dirname,
+      '../fixtures/applitools-ignore-git-merge-base.config.js',
+    );
+    const defaultConfig = {waitBeforeScreenshots: 50};
+    const config = generateConfig({argv: {conf: configPath}, defaultConfig, externalConfigParams});
+
+    // this is the important part, because it's not `true` then `parentBranchBaselineSavedBefore` should be sent in `startInfo`
+    delete config.ignoreGitMergeBase;
+
+    let results = await eyesStorybook({
+      config: {
+        serverUrl,
+        storybookUrl: 'http://localhost:9001',
+        ...config,
+      },
+      logger,
+      performance,
+      timeItAsync,
+      outputStream: stream,
+    });
+    results = flatten(results.map(r => r.resultsOrErr));
+    for (const testResults of results) {
+      const sessionUrl = `${serverUrl}/api/sessions/batches/${encodeURIComponent(
+        testResults.getBatchId(),
+      )}/${encodeURIComponent(testResults.getId())}`;
+
+      const session = await fetch(sessionUrl).then(r => r.json());
+      expect(session.startInfo.parentBranchBaselineSavedBefore).to.match(
+        /\d{4}-\d{2}-\d{2}T\d{2}\:\d{2}:\d{2}\+\d{2}\:\d{2}/,
+      );
+    }
+  });
+
+  it('handles ignoreGitMergeBase', async () => {
+    const {stream} = testStream();
+    const configPath = path.resolve(
+      __dirname,
+      '../fixtures/applitools-ignore-git-merge-base.config.js',
+    );
+    const defaultConfig = {waitBeforeScreenshots: 50};
+    const config = generateConfig({argv: {conf: configPath}, defaultConfig, externalConfigParams});
+
+    let results = await eyesStorybook({
+      config: {
+        serverUrl,
+        storybookUrl: 'http://localhost:9001',
+        ...config,
+      },
+      logger,
+      performance,
+      timeItAsync,
+      outputStream: stream,
+    });
+    results = flatten(results.map(r => r.resultsOrErr));
+    for (const testResults of results) {
+      const sessionUrl = `${serverUrl}/api/sessions/batches/${encodeURIComponent(
+        testResults.getBatchId(),
+      )}/${encodeURIComponent(testResults.getId())}`;
+
+      const session = await fetch(sessionUrl).then(r => r.json());
+      expect(session.startInfo.parentBranchBaselineSavedBefore).to.be.undefined;
+    }
   });
 });