+++ packages/eyes-images/CHANGELOG.md
@@ -3,11 +3,21 @@
 ## Unreleased
 
 
+## 4.19.1 - 2021/3/11
+
+- fix: update getScreenshot to support the correct return type when performing a full page stitch ([Trello](https://trello.com/c/6oVADkSf))
+- updated to @applitools/eyes-sdk-core@12.15.0 (from 12.14.10)
+
+## 4.19.0 - 2021/2/22
+
+- add `extractText` and `extractTextRegions` support
+- updated to @applitools/eyes-sdk-core@12.14.10 (from 12.13.5)
+
 ## 4.18.1 - 2021/1/24
 
 - chore: add husky
 - updated to @applitools/eyes-sdk-core@12.13.5 (from 12.10.0)
-- updated to @applitools/eyes-sdk-core@12.13.5 (from 12.10.0)
+
 ## 4.18.0 - 2020/12/18
 
 - updated to @applitools/eyes-sdk-core@12.10.0 (from 12.9.1)