+++ packages/eyes-testcafe/CHANGELOG.md
@@ -3,6 +3,16 @@
 ## Unreleased
 
 
+## 1.12.7 - 2021/3/15
+
+- updated to @applitools/eyes-sdk-core@12.16.0 (from 12.15.0)
+
+## 1.12.6 - 2021/3/3
+
+- fix bug that prevented window resizing from working reliably ([Trello](https://trello.com/c/xNCZNfPi))
+- updated to @applitools/eyes-sdk-core@12.15.0 (from 12.14.10)
+- updated to @applitools/visual-grid-client@15.6.0 (from 15.5.20)
+
 ## 1.12.5 - 2021/2/2
 
 - fix default checkWindow behavior- take full page when no target is specified